+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## master
 * upgrade to Django 1.10
+* "External" loader that fetches documents from a URL
 
 ## 0.1 (2016-04-28)
 * initial release